requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Google AI integration
google-generativeai==0.3.2

# Additional utilities
pydantic==2.5.0
typing-extensions==4.8.0

# Development dependencies (optional)
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1